Pass data of the correct type to libtopotoolbox's hillshade (#262)
If any of the arrays have a data type other than float32, pybind11 will
silently create a copy and pass it to libtopotoolbox, thus not filling
out the one that we allocated. This change makes sure that everything
is float32 within hillshade.

Adds a test to confirm that hillshade works for float64 DEMs