Account for different node ordering of flowpathextract in its test
flowpathextract now returns topologically sorted flow paths, so the
previous test, which compared the extracted flow path to the stream
array of a StreamObject constructed using only that flow path, would
fail. These two arrays have the same elements, but in different
orders. The test therefore loops over the edge list to reconstruct the
topological ordering of the stream nodes.