Fix memory order sensitivity of `flowpathextract` (#268)
See #199

This PR adds a test to make sure that flowpathextract returns the same flow path regardless of the memory order of the input DEMs.

To fix this test, we thread the memory order through the intermediate arrays. However, the result is still sensitive to the memory order because the return value is sorted according to pixel index. We now return a topologically sorted flow path, which is done by indexing through a new FlowObject.stream array. This is the counterpart of StreamObject.stream, and it contains the linear indices of every pixel in topological order. The topologically sorted return value is now also consistent with the MATLAB implementation, which constructs the flow path by iterating over the edge list.

This change breaks the existing flowpathextract test, because the flowpathextract return value is now topologically sorted while the StreamObject.stream array is not. Since the latter is not topologically sorted in the MATLAB implementation, I left that alone and instead modified the flowpathextract test to reconstruct the topological order of the stream array.