Hide the memory ordering from users of StreamObject (#276)
The memory ordering is an implementation detail that users really
shouldn't have to worry about. This change makes the
StreamObject.order attribute "private" by renaming it to
StreamObject._order and provides a few accessor functions for
obtaining the indices of nodes and edges in the stream network.

Since numpy does not perform linear indexing, we need to unravel the
linear indices before indexing into an array like a GridObject. The
accessors do this unraveling and return a tuple of two arrays, one
with the row indices, one with the column indices. This tuple can be
used to directly index into an array.

The provided accessors are

- StreamObject.node_indices :: the indices of the nodes in the stream
network. This is a node attribute list, and it is a property, so it is
accessed without the parentheses

- StreamObject.node_indices_where(nal) :: the indices filtered by the
Boolean node attribute list. This is a function. It allows you to
retrieve the indices of subsets of the nodes, such as outlets, using
something like s.node_indices_where(s.streampoi('outlets'))

- StreamObject.source_indices :: the indices of the sources of each
edge. This is an edge attribute list and a property.

- StreamObject.target_indices :: the indices of the targets of each
edge. This is an edge attribute list and a property.

The unravel_index calls have largely been removed from
stream_object.py and test_stream_object.py.