Add multidimensional index accessors to FlowObject (#277)
This PR mirrors #276 for FlowObject and more or less resolves #202.

The source_indices and target_indices properties work just as they do for StreamObject. node_indices and node_indices_with from StreamObject are not included because the node indices are just the indices of the GridObject in topological order, and I have not found an application of these yet.

There is also a FlowObject.unravel_index method that takes linear indices from the user and unravels them according to the shape and memory order of the FlowObject. This is just a wrapper around Numpy's unravel_index function. I'm not sure that it is exactly necessary, but there are several places in the FlowObject methods and in the tests that this is necessary. I did not implement the inverse, ravel_multi_index, because there are only a few locations in the tests where this is used.