Factor out the StreamObject isomorphism tests
The isequivalent test uses the new node_indices, source_indices, and
target_indices to create memory-order independent sets of nodes and
edges for two StreamObjects. and compares them. This change replaces
the idxmap and comparisons in many of the memory order tests. It also
finishes hiding the `order` parameter of FlowObject within the
tests (see #202).