Modify the order_dems transforms
These are

13   0  0
0  -13  0

which seems to work better with reproject. I will admit that I do not
fully understand the relationship between these or why, for example,
the fdem transform should be the same as the cdem transform, but it
seems to work, even in the hillshade tests which required this in the
first place.