Add a memory order test for GridObject.reproject (#288)
* Add a memory order test for GridObject.reproject

See #199

This test fails on the flag assertions.

* Modify the order_dems transforms

These are

13   0  0
0  -13  0

which seems to work better with reproject. I will admit that I do not
fully understand the relationship between these or why, for example,
the fdem transform should be the same as the cdem transform, but it
seems to work, even in the hillshade tests which required this in the
first place.

* Make sure that reproject returns a similar array

This should make sure that the dtype and order of the input are
preserved by reproject.