Add a memory order test for StreamObject when supplied with explicit stream pixels (#289)
* Add a memory order test for StreamObject with stream_pixels

See #199

This test provides the accumulation threshold directly as an array to
ensure that the StreamObject constructor handles the provided array
correctly.

This fixes the breaking memory order test by replacing
ravel(order='F') with ravel(order='K'). It also takes the opportunity
to remove the `isinstance` checks, replacing them with a single call
to asarray.