Use node_indices_where to supply tuple of channel head indices
I believe this is a better API because it hides the linear indices and
thus the memory ordering from the user. The user provides a tuple of
indices `([r1,r2,...], [c1,c2,...])` which contain the pixel locations
that contain the channel heads, which is more easily accessible.

This test will fail because the channelheads constructor has not yet
been updated to take the tuple of indices.