Make the channelheads constructor of StreamObject order agnostic
Now we pass in a tuple of multidimensional indices for the channel
heads instead of the linear indices, and we construct the influence
map with the correct memory ordering.

This is a breaking change to the StreamObject constructor API because
it requires the 2D indices rather than the linear indices.