Add a memory order test of the StreamObject channelheads constructor (#291)
See #199

This required some modification of the channelheads API: you now have to pass a tuple of arrays as in StreamObject(fd, channelheads=(rows, cols)) instead of the linear indices. This prevents the user having to obtain the linear index of a given pixel, which requires knowledge of the underlying memory order of the StreamObject.

If you have a logical node attribute list, such as is returned by ch = s.streampoi('channelheads'), you can do something like StreamObject(fd, channelheads=s.node_indices_where(ch))

This is a breaking API change to StreamObject.

Other tests were modified to use this new version of the constructor.