Replace shape checks with `validate_alignment` (#301)
Resolves #212 

This set of changes largely turns checks like

```python
if a.shape != b.shape:
```

into the equivalent

```python
if not validate_alignment(a, b):
```

This is equivalent, but it also checks the georeferencing information and hides some more of the internal details of the various objects.

`validate_alignment` had to be moved to a separate module to avoid circular imports. This module is called `interface` for now, because it specifies an interface to TopoToolbox objects.

Some of the changes are accompanied by changes that use `ezgetnal` or `asarray` to avoid excessive duplication.