Add resize and zscore (#278)
Closes #264

This adds the functions zscore and resize to the GridObject.

Maybe crop would be a more fitting name for the resize function. Also, having the arguments left, right, top and bottom might be unnecessary. The function only needs two x and y values each and figures out which is left/right or top/bottom automatically. Using highlight_selected will create a plot which highlights the area the new GridObject will be cropped to.
I am not 100% sure the updating of the bounds and transform is completely correct, so if anything seems off about that please let me know.