Crslin (#307)
This PR adds a crslin method to StreamObject that implements constrained regularized smoothing with a least squares objective function. The kvxopt package is used for solving the resulting quadratic program. kvxopt is API-equivalent to cvxopt, but is used here because `kvxopt` provides compiled binary wheels on a wider variety of platforms. This dependency should be reviewed in the future to ensure that it continues to meet our needs,.

A test is added that ensures the resulting modeled stream profile obeys the constraints.