Implement lowerenv
Resolves #305
Resolves #306

This PR implements lowerenv, which computes the lower convex envelope
of a stream profile. It relies on the libtopotoolbox function of the
same name (TopoToolbox/libtopotoolbox#196).

CMakeLists.txt is updated with the tag of libtopotoolbox that includes
lowerenv.

src/lib/stream.cpp wraps lowerenv.

src/topotoolbox/stream_object.py implements lowerenv as a method on
StreamObject.

tests/test_stream_object.py adds a property-based test of convexity of
the profile. This is the same property-based test used in the
libtopotoolbox test suite. It also adds a memory ordering test (#199).

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>