Prototype a knickpoint finder
See #304

This is the start of the Python implementation of knickpointfinder. It
is currently a loop around `lowerenv` (#304, #310) that adds
knickpoints that are above a certain tolerance.

Discrepancies between this and the MATLAB version include:

1. MATLAB TopoToolbox checks reaches of the stream network between
knickpoints simultaneously, so it can identify more than one
knickpoint at a time. It uses dmperm to compute these connected
components. We could do something like this with some clever
traversals.

2. The return value is a logical node attribute list that is True for
all knickpoints. All of the other information like the x and y
coordinates and upstream distances can be reconstructed using the node
attribute list.

3. No plotting or intermediate output is given.

4. It is not possible to initialize the knickpointfinder with knickpoints.

5. Python and MATLAB do not identify the same knickpoints. Different
numbers of knickpoints can be identified for the same tolerance, and
different points are identified. It is currently unclear whether these
are because of differences in flow routing, (1) above or numerical
discrepancies in lowerenv.