Compute multiple knickpoints at a time by traversing over reaches
The maximum difference between the profile and the lower convex
envelope is tracked downstream, observing breaks at knickpoints, and
then tracked back upstream, observing breaks at knickpoints. The
indices of the maximum are also tracked, and the unique values
represent potential knickpoints in each reach. Any of these that are
greater than the provided tolerance are identified as knickpoints in
each loop.