Implement conncomps (#314)
This is very similar to computing drainage basins: we label each
component at the outlet and then propagate those labels upstream.

test_conncomps ensures that each node has the same label as its
downstream neighbor and checks that the labels are reasonable. Note
that we have to clean the StreamObject to ensure that there are no
single-pixel networks, which don't get labeled properly.

test_conncomps_order is a standard memory order test (#199), but it
does have to account for differences in the labels. We do this in our
FlowObject drainage basins test, and this code is copied from there.