Replace uses of np.bool with bool (#315)
This avoids deprecation warnings when testing on older Numpy versions.

numpy.bool as an alias for Python's built-in bool was deprecated in
Numpy v1.20 while Numpy's numpy.bool_ type was renamed to numpy.bool
in v2.0. All of the options seem to produce identical arrays using the
numpy.bool type, and this change helps avoid the deprecation warnings,
which make the test logs quite noisy.