Implement `knickpointfinder` (#316)
See #304

This is the start of the Python implementation of knickpointfinder. It
is currently a loop around `lowerenv` (#304, #310) that adds
knickpoints that are above a certain tolerance.

Discrepancies between this and the MATLAB version include:

1. MATLAB TopoToolbox checks reaches of the stream network between
knickpoints simultaneously, so it can identify more than one
knickpoint at a time. It uses dmperm to compute these connected
components. We could do something like this with some clever
traversals.

2. The return value is a logical node attribute list that is True for
all knickpoints. All of the other information like the x and y
coordinates and upstream distances can be reconstructed using the node
attribute list.

3. No plotting or intermediate output is given.

4. It is not possible to initialize the knickpointfinder with knickpoints.

5. Python and MATLAB do not identify the same knickpoints. Different
numbers of knickpoints can be identified for the same tolerance, and
different points are identified. It is currently unclear whether these
are because of differences in flow routing, (1) above or numerical
discrepancies in lowerenv.

* Pass initial knickpoints to the knick point finder

* Copy the provided knickpoints array

We need to make a copy to prevent overwriting the initial knickpoints
array.

* Add a few tests of the knickpoint finder

The first runs knickpointfinder with some preselected knickpoints,
but does no iterations, so it should find no additional
knickpoints.

The second runs knickpointfinder for one iteration with no tolerance,
so only one knickpoint should be identified.

* Test that the knickpoints array is properly copied

* Compute multiple knickpoints at a time by traversing over reaches

The maximum difference between the profile and the lower convex
envelope is tracked downstream, observing breaks at knickpoints, and
then tracked back upstream, observing breaks at knickpoints. The
indices of the maximum are also tracked, and the unique values
represent potential knickpoints in each reach. Any of these that are
greater than the provided tolerance are identified as knickpoints in
each loop.

* Add some explanatory comments

* Update documentation

* Fix kpcopy test

Since we identify multiple knickpoints on a single iteration, this
assertion is no longer true.

* Remove extraneous assignment to dz that trips mypy type checks