Call autoscale_view automatically from StreamObject.{plot, plotdz}
Resolves #335

This change calls autoscale_view from inside the plot and plotdz
methods, so users don't have to call it externally. It adds two
keyword arguments, scalex and scaley, that are passed to
Axes.autoscale_view() and can be used to control whether the x or y
axis is automatically scaled. This could be helpful if you have
already plotted a dataset that spans a subset of the stream network
and would like to plot the stream network over that dataset without
zooming out to the entire stream.

Alternatively, setting the axis limits manually with set_xlim and
set_ylim turns autoscaling off for that axis, so later calls to
autoscale_view will not automatically scale the axis, regardless of
the values of scalex and scaley. If a user has explicitly constrained
the axis limits, this change will not surprise them.

Documentation has been updated, including the StreamObject tutorial
notebook.