Use hillshade_fused by default
This is the lower-memory version of hillshade available from
libtopotoolbox. It can be slightly faster for large DEMs and avoids
excessive memory allocation. An option is provided to hillshade to
used the fused implementation or not, with the default to use the
fused version.

A test is added to ensure that the fused and unfused implementations
return the same values.